Package name changed from example to reft
As the default package name is com.example which is not allowed for
apps, changed the package to com.reft. Also Nearest gas station json
parsing and populating to listview is over. Latest APK will support both
portrait and landscape views. Have uploaded the same for Alpha Testing
in Google Play Store :+1: